+++ src/minified-web-full-src.js
@@ -1247,9 +1247,9 @@ define('minified', function() {
 
 		if (element) {
 			if (isString(spec)) {
-				var name = replace(replace(spec, /^%/, 'data-'), /^[$@]+/);
+				var name = replace(replace(replace(spec, /^\$float$/, 'cssFloat'), /^%/, 'data-'), /^[$@]+/);
 				var s;
-				if (spec == '$')
+				if (spec == '$') 
 					s = element.className;
 				else if (spec == '$$') {
 					// @condblock ie8compatibility
@@ -1474,7 +1474,7 @@ define('minified', function() {
     		 else
     			// @condend fadeslide
     			 flexiEach(self, function(obj, c) {
-    				 var nameClean = replace(replace(name, /^%/,'data-'), /^[@$]+/);
+    				 var nameClean = replace(replace(replace(name, /^\$float$/, 'cssFloat'), /^%/,'data-'), /^[@$]+/);
     				 var className = obj['className'] || '';
     				 var newObj = /^\$/.test(name) ? obj.style : obj;
     				 var newValue = isFunction(value) ? value($(obj).get(name), c, obj) : value;