+++ CHANGES.md
@@ -5,10 +5,12 @@ Beta 5 (in progress)
 --------------------
 * added up()
 * added not()
+* added isEmpty()
 * formatNumber() supports any character as group separator
 * formatNumber() will limit the number of digits in front of the decimal separator, unless you use a single '#'
 * updated Minified's promise implementation to the latest Promises/A+ spec (full package only)
 * a promise's state can now be obtained by invoking it as a function without arguments (see promise(), full package only)
+* set() and get() will automatically convert $float to $cssFloat
 
 Beta 4.2 (2014-12-26)
 ---------------------